+++ geoserver_manager/toolbelt/preferences.py
@@ -21,28 +21,9 @@ from geoserver_manager.toolbelt.env_var_parser import EnvVarParser
 # ########## Classes ###############
 # ##################################
 
+# Every setting can be overridden by an environment variable of this prefix
+# and its upper-cased name: QGIS_GEOSERVER_MANAGER_DEBUG_MODE=true.
 PREFIX_ENV_VARIABLE = "QGIS_GEOSERVER_MANAGER_"
-
-
-@dataclass
-class PlgEnvVariableSettings:
-    """Plugin settings from environnement variable"""
-
-    def env_variable_used(self, attribute: str, default_from_name: bool = True) -> str:
-        """Get environnement variable used for environnement variable settings
-
-        :param attribute: attribute to check
-        :type attribute: str
-        :param default_from_name: define default environnement value from attribute name PREFIX_ENV_VARIABLE_<upper case attribute>
-        :type default_from_name: bool
-        :return: environnement variable used
-        :rtype: str
-        """
-        settings_env_variable = asdict(self)
-        env_variable = settings_env_variable.get(attribute, "")
-        if not env_variable and default_from_name:
-            env_variable = f"{PREFIX_ENV_VARIABLE}{attribute}".upper()
-        return env_variable
 
 
 @dataclass
@@ -133,28 +114,18 @@ class PlgOptionsManager:
         :return: plugin settings
         :rtype: PlgSettingsStructure
         """
-        # get dataclass fields definition
-        settings_fields = fields(PlgSettingsStructure)
-        env_variable_settings = PlgEnvVariableSettings()
-
         # retrieve settings from QGIS/Qt
         settings = QgsSettings()
         settings.beginGroup(__title__)
 
-        # map settings values to preferences object
-        li_settings_values = []
-        for i in settings_fields:
-            try:
-                value = settings.value(key=i.name, defaultValue=i.default, type=i.type)
-                # If environnement variable used, get value from environnement variable
-                env_variable = env_variable_settings.env_variable_used(i.name)
-                if env_variable:
-                    value = EnvVarParser.get_env_var(env_variable, value)
-                li_settings_values.append(value)
-            except TypeError:
-                li_settings_values.append(
-                    settings.value(key=i.name, defaultValue=i.default)
-                )
+        # map settings values to preferences object, the environment winning
+        li_settings_values = [
+            EnvVarParser.get_env_var(
+                f"{PREFIX_ENV_VARIABLE}{i.name}".upper(),
+                settings.value(key=i.name, defaultValue=i.default, type=i.type),
+            )
+            for i in fields(PlgSettingsStructure)
+        ]
 
         # instanciate new settings object
         options = PlgSettingsStructure(*li_settings_values)